currently keas.build will not use unchanged trunk packages if the related branch package was released. This ends in pick a branch package release version for the next trunk release. NOTE: this is fine and correct as long as the branch contains bugfixes. But our branch contains a new complete product rewrite (new product generation). Such a new product rewrite contains the same packages but complete new software which could not get mixed within trunk package.
NOTE: I guess the above is not really correct. Imagine what happens if you change a package in the trunk where you have a bugfix
release for from a branch and didn't backport the bugfix from the branch to the trunk. Such bugfixes get lost because the next
release will get released from the trunk. (the trunk changes forces to do so). But that's another problem and has nothing to do 
with what I'm try to solve. (that's just a bugfix must get back ported ASAP problem)

What I try to do now is to find a way that a trunk/branch (option -b branch in keas.build) package release get not mixed within other branches. If this is not possible the fallback would be to make sure that all branches, where release get made from, get involved in find a new version. This would at least make sure that we pick a newer version then ever was released from any trunk or branch for each package.

In short words, we need completly independent trunk and branche releases because our branch contains a new software generation which is not compatible with the trunk. Currently keas.build only uses branch release as future package fixes for the trunk.